send all pgn each push
