+++ package.json
@@ -15,36 +15,36 @@
   "dependencies": {
     "@headlessui/vue": "^1.7.23",
     "@heroicons/vue": "^2.2.0",
-    "@tauri-apps/api": "^2.2.0",
+    "@tauri-apps/api": "^2.3.0",
     "@tauri-apps/plugin-dialog": "~2.2.0",
     "@tauri-apps/plugin-fs": "~2.2.0",
-    "@tauri-apps/plugin-notification": "~2.2.1",
+    "@tauri-apps/plugin-notification": "~2.2.2",
     "@tauri-apps/plugin-shell": "~2.2.0",
-    "@tauri-apps/plugin-updater": "~2.5.0",
+    "@tauri-apps/plugin-updater": "~2.5.1",
     "debounce": "^2.2.0",
     "fetch-ndjson": "^1.1.7",
-    "openapi-fetch": "^0.13.4",
+    "openapi-fetch": "^0.13.5",
     "pinia": "^3.0.1",
     "pinia-plugin-persistedstate": "^4.2.0",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   },
   "devDependencies": {
-    "@faker-js/faker": "^9.5.0",
-    "@lichess-org/types": "^2.0.27",
-    "@tailwindcss/postcss": "^4.0.6",
-    "@tauri-apps/cli": "^2.2.7",
-    "@types/node": "^22.13.4",
+    "@faker-js/faker": "^9.6.0",
+    "@lichess-org/types": "^2.0.28",
+    "@tailwindcss/postcss": "^4.0.14",
+    "@tauri-apps/cli": "^2.3.1",
+    "@types/node": "^22.13.10",
     "@vitejs/plugin-vue": "^5.2.1",
-    "chess.js": "1.0.0",
+    "chess.js": "1.1.0",
     "jsdom": "^26.0.0",
-    "postcss": "^8.5.2",
-    "prettier": "^3.5.1",
-    "tailwindcss": "^4.0.6",
-    "typescript": "^5.7.3",
-    "vite": "^6.1.0",
-    "vitest": "^3.0.5",
-    "vue-tsc": "^2.2.2"
+    "postcss": "^8.5.3",
+    "prettier": "^3.5.3",
+    "tailwindcss": "^4.0.14",
+    "typescript": "^5.8.2",
+    "vite": "^6.2.2",
+    "vitest": "^3.0.8",
+    "vue-tsc": "^2.2.8"
   },
-  "packageManager": "pnpm@10.4.1+sha512.c753b6c3ad7afa13af388fa6d808035a008e30ea9993f58c6663e2bc5ff21679aa834db094987129aa4d488b86df57f7b634981b2f827cdcacc698cc0cfb88af"
+  "packageManager": "pnpm@10.6.3+sha512.bb45e34d50a9a76e858a95837301bfb6bd6d35aea2c5d52094fa497a467c43f5c440103ce2511e9e0a2f89c3d6071baac3358fc68ac6fb75e2ceb3d2736065e6"
 }
