+++ src-tauri/Cargo.toml
@@ -17,8 +17,6 @@ askama = "0.12.1"
 fix-path-env = { git = "https://github.com/tauri-apps/fix-path-env-rs" }
 oauth2 = "4.4.2"
 open = "5.0.1"
-reqwest = { version = "0.11.23", features = ["blocking", "json"] }
-serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.111"
 serde_urlencoded = "0.7.1"
 tauri = { version = "2", features = [ "wry", "objc-exception", "devtools", "native-tls-vendored"], default-features = false }
