+++ src/App.vue
@@ -1,10 +1,9 @@
 <script setup lang="ts">
 import { useUserStore } from './stores/user';
 import { listen } from '@tauri-apps/api/event';
-import { AccessTokenResponse, PgnPushResult } from './types';
+import { AccessTokenResponse } from './types';
 import { useLogStore } from './stores/logs';
 import { requestNotificationPermission } from './notify';
-import { pgnTag } from './utils';
 import { useFavoritesStore } from './stores/favorites';
 
 const logs = useLogStore();
@@ -14,27 +13,6 @@ const favorites = useFavoritesStore();
 listen<AccessTokenResponse>('event::update_access_token', event => {
   logs.clear();
   user.setAccessToken(event.payload);
-});
-
-listen<number>('event::queue_size', event => {
-  logs.queueSize = event.payload;
-});
-
-listen<PgnPushResult>('event::upload_success', event => {
-  event.payload.files.forEach(file => {
-    logs.files.add(file);
-  });
-
-  logs.info(`Uploaded ${event.payload.files.join(', ')}`);
-
-  const errors = event.payload.response.games.filter(game => game.error);
-  errors.forEach(game => {
-    logs.error(`PGN Error: ${game.error} in ${pgnTag('White', game.tags)} vs ${pgnTag('Black', game.tags)}`);
-  });
-});
-
-listen<string>('event::upload_error', event => {
-  logs.error(event.payload);
 });
 
 if (user.isLoggedIn()) {
