+++ tsconfig.json
@@ -23,6 +23,6 @@
 
     "types": ["vitest/importMeta"]
   },
-  "include": ["src/**/*.ts", "sample-data/**/*.ts", "src/**/*.tsx", "src/**/*.vue"],
+  "include": ["src/**/*.ts", "sample-data/**/*.ts", "src/**/*.tsx", "src/**/*.vue", "tests/**/*.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
